Convert command ssh to oclif.
Change-type: patch
