+++ lib/app-capitano.js
@@ -72,9 +72,6 @@ capitano.command(actions.tunnel.tunnel);
 // ---------- Preload Module ----------
 capitano.command(actions.preload);
 
-// ---------- SSH Module ----------
-capitano.command(actions.ssh.ssh);
-
 // ---------- Local balenaOS Module ----------
 capitano.command(actions.local.configure);
 capitano.command(actions.local.flash);
